Move fitness to own file
